+++ ios/Classes/EmmaSerializer.swift
@@ -14,7 +14,7 @@ class EmmaSerializer {
         return [
             "id": nativeAd.idPromo,
             "templateId": nativeAd.nativeAdTemplateId ?? "",
-            "cta": nativeAd.getField("CTA") ?? "",
+            "cta": nativeAd.nativeAdContent["CTA"] ?? "",
             "times": nativeAd.times,
             "tag": nativeAd.tag ?? "",
             "params": nativeAd.params ?? [:],
@@ -29,6 +29,10 @@ class EmmaSerializer {
                 return .Startview
             case "nativeAd":
                 return .NativeAd
+            case "adBall":
+                return .Adball
+            case "strip":
+                return .Strip
             default:
                 return nil
         }
@@ -40,6 +44,10 @@ class EmmaSerializer {
             return EMMACampaignType.campaignStartView
         case InAppType.NativeAd:
             return EMMACampaignType.campaignNativeAd
+        case InAppType.Adball:
+            return EMMACampaignType.campaignAdBall
+        case InAppType.Strip:
+            return EMMACampaignType.campaignStrip
         default:
             return nil
         }